Feature_EnemyHpBar
적 체력에 대한 UI 추가

+++ Assets/Scripts/Enemy/EnemySO.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Enemy", menuName = "Characters/Enemy")]
 public class EnemySO : ScriptableObject
 {
-    [SerializeField] public int maxHealth;
+    [SerializeField] public float maxHealth;
     [SerializeField] public int Damage;
     [SerializeField] public float evasionRate;
 }
